requirements.txt
Django==5.2.7
channels>=4.1.0,<5.0.0
daphne>=4.1.0,<5.0.0
python-telegram-bot>=20.7,<21.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
# On Windows, provide time zone data package used by Django
tzdata>=2024.1; sys_platform == "win32"
